requirements.txt
requests
pillow
numpy
opencv-python
pyproj